requirements.txt
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==43.0.0
dill==0.3.8
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-redis==5.4.0
gunicorn==23.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
pylint==3.2.6
python-decouple==3.8
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
